Ajustado controller playoffController.ts
